requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.6
